CliSignerPKCS7

+++ src/Esia/Signer/CliSignerPKCS7.php
@@ -4,32 +4,10 @@ namespace Esia\Signer;
 
 use Esia\Signer\Exceptions\SignFailException;
 use Psr\Log\LoggerAwareTrait;
-use Psr\Log\NullLogger;
-
 
 class CliSignerPKCS7 extends AbstractSignerPKCS7 implements SignerInterface
 {
     use LoggerAwareTrait;
-
-    /**
-     * SignerPKCS7 constructor.
-     * @param string $certPath
-     * @param string $privateKeyPath
-     * @param string $privateKeyPassword
-     * @param string $tmpPath
-     */
-    public function __construct(
-        string $certPath,
-        string $privateKeyPath,
-        ?string $privateKeyPassword,
-        string $tmpPath
-    ) {
-        $this->certPath = $certPath;
-        $this->privateKeyPath = $privateKeyPath;
-        $this->privateKeyPassword = $privateKeyPassword;
-        $this->tmpPath = $tmpPath;
-        $this->logger = new NullLogger();
-    }
 
     /**
      * @param string $message
@@ -37,7 +15,8 @@ class CliSignerPKCS7 extends AbstractSignerPKCS7 implements SignerInterface
      * @throws SignFailException
      */
 
-    public function sign(string $message): string {
+    public function sign(string $message): string
+    {
         $this->checkFilesExists();
 
         // random unique directories for sign
@@ -46,16 +25,21 @@ class CliSignerPKCS7 extends AbstractSignerPKCS7 implements SignerInterface
         file_put_contents($messageFile, $message);
 
         $this->run(
-            'openssl '.
-            'smime -sign -binary -outform DER -noattr '.
-            '-signer '.escapeshellarg($this->certPath).' '.
-            '-inkey '.escapeshellarg($this->privateKeyPath). ' '.
-            '-passin '.escapeshellarg('pass:'.$this->privateKeyPassword).' '.
-            '-in '.escapeshellarg($messageFile).' '.
-            '-out '.escapeshellarg($signFile)
+            'openssl ' .
+            'smime -sign -binary -outform DER -noattr ' .
+            '-signer ' . escapeshellarg($this->certPath) . ' ' .
+            '-inkey ' . escapeshellarg($this->privateKeyPath) . ' ' .
+            '-passin ' . escapeshellarg('pass:' . $this->privateKeyPassword) . ' ' .
+            '-in ' . escapeshellarg($messageFile) . ' ' .
+            '-out ' . escapeshellarg($signFile)
         );
 
         $signed = file_get_contents($signFile);
+        if ($signed === false) {
+            $message = sprintf('cannot read %s file', $signFile);
+            $this->logger->error($message);
+            throw new SignFailException($message);
+        }
         $sign = $this->urlSafe(base64_encode($signed));
 
         unlink($signFile);
@@ -63,22 +47,34 @@ class CliSignerPKCS7 extends AbstractSignerPKCS7 implements SignerInterface
         return $sign;
     }
 
-    private function run($command) {
-        $process = proc_open($command, [
-            ['pipe', 'w'], // stdout
-            ['pipe', 'w'], // stderr
-        ], $pipes);
+    /**
+     * @param $command
+     * @return void
+     * @throws SignFailException
+     */
+    private function run(string $command): void
+    {
+        $process = proc_open(
+            $command,
+            [
+                ['pipe', 'w'], // stdout
+                ['pipe', 'w'], // stderr
+            ],
+            $pipes
+        );
+
         $result = stream_get_contents($pipes[0]);
         fclose($pipes[0]);
+
         $errors = stream_get_contents($pipes[1]);
         fclose($pipes[1]);
+
         $code = proc_close($process);
-        if (0 != $code) {
-            $errors = trim($errors) ?: 'unknown';
+        if (0 !== $code || $result === false) {
+            $errors = $errors ?: 'unknown';
             $this->logger->error('Sign fail');
             $this->logger->error('SSL error: ' . $errors);
             throw new SignFailException($errors);
         }
-        return $result;
     }
 }
